requirements.txt
fastparquet
python-snappy
psycopg2-binary
cryptography==3.4.7
boto3==1.17.54
azure-common==1.1.27
azure-core==1.13.0
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-cosmos==4.2.0
apache-airflow-backport-providers-microsoft-mssql==2020.11.23
apache-airflow-backport-providers-odbc==2021.3.17
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-snowflake==1.2.0
apache-airflow-providers-sqlite==1.0.2